agents/support/requirements.txt
fastapi
pydantic
uvicorn
openai
langchain
langchain-openai
python-dotenv
tiktoken
langchain_community 